Unconditionally clone contrail-kolla-ansible
The contrail-kolla-ansible repo needs to be cloned even though the orchestrator
is set to k8s because of the static nature of ansible playbook includes. This
will cause errors when running the install_contrail playbook. The playbooks from
contrail-kolla-ansible will be skipped, but need to be present when running with
orchestrator=k8s.

The ideal solution is to have a separate playbook to install/run contrail with
openstack. Shall do this in an incremental commit. Currently doing first option
to unblock k8s folks

Change-Id: I1b1be85ee9d0bbdc8b9e1506e0a05473c790c59b
Partial-Bug: #1752219