Fix deriving network_interface based on CONTROL_DATA_NETLIST
Side effect of bug #1755590
The network_interface_detected variable should be per-host and cannot be global.
Similarly for external_interface_detected

PATCHSET 4: Accomodate the case where "roles" is not defined at all for an
instance defined in instances.yaml

PATCHSET 5: When it is not possible to detect network_interface, use the
management network interface

PATCHSET 6: Handle the case where hostvars may not contain all the nodes in the
cluster. This can happen when the openstack roles are not colocated with
contrail roles.

Change-Id: I4939611187c39fa216e455ebd35441d5b789be7b
Closes-Bug: #1759731