Closes-Bug: #1760051 - Creation of SSL certs per Server for Metadata SSL
As in SM, these certs are created for each server and Subject Alt Names are set based on IP/DNS
To enable SSL, please add to config/instances.yml:

contrail_configuration:
  SSL_ENABLE: True

This will create and copy the certs to the "default" paths specified here:
https://github.com/Juniper/contrail-container-builder/blob/master/containers/base/common.sh#L26

We will need to implement separately what to do if user created certs are supplied.
This ability is not there in container-builder today so I haven't implemented it.

Change-Id: I322b153e46c3a86119626f3ec8315a31aad2e4e5