bug #1773103
Allow user to define more than one interfaces for kvm. Before this change only eth0 was configured with provided ip and bridge, another interface was attached to the vm but not configured.

In most cases, customer requires to split mgmt, control and data plane. Also, vRouters may require adding additional interface when it plays as gateway mode.

This change allow the customer to define how many additional interfaces should be configured and also if an ip/mask/gateway/dns should be configured in each interface.

Following is the new syntax supported in kvm for an instance. Notice the section provided in additional_interfaces. eth0 will be configured as usual and the 3 additional interfaces provided (eth1, eth2 and eth3) will be configured as defined by the user. eth1 with ip/mask and eth2 and eth3 without ip/mask.

  contrailvrouter1:
    provider: kvm
    host: 127.0.0.1
    bridge: b33mgmt
    ip: 192.168.33.41
    additional_interfaces:
      - bridge: b34data
        ip: 192.168.34.41
        mask: 255.255.255.0
      - bridge: b35ext
      - bridge: b36ext
    roles:
        vrouter:
          PHYSICAL_INTERFACE: eth1
        k8s_node:

Note: the change requires using index while looping in ansible. Given that contrail-ansible 2.4 version is required then "with_indexed_items" has been used. After moving to ansible 2.5 this should be updated to use loop: and loop_control: index_var options which is the new way in ansible 5. Also, this will require to change additional_interface.j2 accordantly.

Change-Id: I63be0c4157b8f82f0fea02a42df6e48bd1aeeb91
Signed-off-by: joncastro <castro.jon@gmail.com>