Support stats collector destination in ansible
Stat messages are written to socket file which will
be read by stats collector. The socket file to be
written to is specified in instances.yaml. The folder
holfing the socket file will be mounted by all the
containers. The user specified STATS_COLLECTOR_DESTINATION_PATH
in the instances.yaml.

Change-Id: I1231c4d7e3475721d6eff75d461e34968c5ace38
Depends-On: I15134efc1015139865730747210cde7160520a05
Partial-Bug: 1794107