Support stats collector destination in ansible
Stat messages are written to socket file which will
be read by appformix and stored. The socket file to be
written to is specified in instances.yaml. The folder
holfing the socket file will be mounted by all the
containers. The user specified STATS_COLLECTOR_DESTINATION_PATH
in the instances.yaml. The defaule value if not
specified is '/var/log/contrail'. The reason for making a
default value if left unspecified is to avoid conditional
checks in the yaml file. There is another variable which
will be derived from this value namely, STATS_COLLECTOR_PATH
which will be used by the container builders to specify
the actual socket path.

Change-Id: I1231c4d7e3475721d6eff75d461e34968c5ace38
Partial-Bug: 1794107