Remove incorrect exception handling in calculate_contrail_roles.py
There are two places, where possible error is actually terminal.
Currently exception caused by failure in HTTP request is changed
to special return value that is not handled anywhere and causes
exception later in Ansible. That exception is difficult to track
and error message is actually meaningless. It's better to let
original exception cause a crash - error message is easy to
understand and clearly points to the cause of failure.

Closes JIRA bug: JCB-219041

Change-Id: Ie65a7b03ffc498aadd1559cf48e06bdf07ea12e5