More optimizations for scaling
1. /etc/hosts entries are now added only on config_db or on openstack_controller
nodes as this is required only for rabbitmq to work. Note: It can be further
optimized to include entries for only the nodes where rabbitmq is going to be
run, but currently it will include entries for all nodes in cluster.

2. package installation transactions are minimized when packages are given as a
list rather than using "with_items" loop

3. Turn of display_skipped_hosts in ansible.cfg. This seems to speed things up
when there are lots of skipped nodes like when generating /etc/hosts for 100+
nodes

Change-Id: I2610e69f1ca53a4ed12b1c6a5acb46103460d650
Partial-Jira-Bug: JCB-219245