Optimize when manage_etc_hosts is 'false'
In scaled cluster of 500 nodes, even when manage_etc_hosts is set to
false, the tasks to change /etc/host entries take more than 2 hours to
skip. This is because of the way ansible evaluates task conditionals
before deciding to skip a task. To avoid this, all the tasks are moved
into a separate file and conditionally included.

Change-Id: I818620bdd4f57ee9e92ecec5973ad5aa43e032aa
Closes-Jira-Bug: JCB-219245