Fix deletion of computes from cluster
Change default value of ENABLE_DESTROY to True

Having this set to False by default prevents the contrail_command UI from
being able to delete compute nodes. If there is a need to run the
playbooks for batches of computes, then the ENABLE_DESTROY needs to be
set to False explicitly in the instances.yaml.

1. Use the instance_name that is registered with config_api when
deleting a compute - not just hostname to allow for fqdn changes.
2. deregister tasks were being called twice. Remove the delete being
called from inside the contrail role.
3. to detect a reprovision the filter plugin queries for the openstack
endpoints URL. This needs to be done using the auth token to work.

TODO:
1. Make it work for rocky
2. Test deleting TSN node
3. Test keystone v3
4. Test SSL_ENABLE set to true
5. Test with enable_haproxy enabled

Change-Id: I3a9d9a66ae89938fdc7d059ca1099dba0203be87
Partial-Jira-Bug: CEM-5141