Detect existing openstack computes to delete
Use the public openstack endpoint to detect existing computes as the
internal IP may not be accessible from the deployer node.

If TLS is enabled, use https as protocol.

In case where openstack service runs on ctrl-data network, the query for
os-hypervisors will return only ctrl-data IP addresses, we need to
figure out the appropriate instance key by going through all the IP
addresses for all the hosts and then matching the IP that matches the
'ip' field of the instances dictionary. This is all done inside the
filter plugin code

TODO: Move the filter plugins for openstack_deployer and
contrail_deployer into a common place so the code can be shared. There
is too much replication currently

Change-Id: I3e903e7bd80442fd67fd0284f3a4121e390c1a67
Partial-Jira-Bug: CEM-5141