Removal of specifying docker_images_tar from UI.
UI will not set this value.
It will be in the format of
/opt/software/xflow/xflow-<version>.tar.gz and ansible will be
/opt/software/xflow/xflow-ansible-<version>.tar.gz

version needs to be specified in config.yml
This makes sure that user should not be able install any version of xflow with any version of CC, else
there may be some compatibility issue which user does not want to see.

closes-jira-bug: CEM-9020
Change-Id: I847c55e61a1733a145c6181e21ecdd80312ffe29