Removed a race condition in some cases
This is related to an ocassional case in Multicloud deployment, where vHost can come up too early and block communication with the Controller until the setup is finished. In that case, any non-localhost tasks will be blocked for a long time then
fail.

This fix addresses the issue by ensuring launching vRouter is the last task in the role.

Closes-Jira-Bug: CEM-9459
Change-Id: I87874a610b9f19d00106dbaf4a3500b12a11b170